Cleanup done.
