+++ go.mod
@@ -3,12 +3,20 @@ module github.com/scottjbarr/pricebroadcaster
 go 1.18
 
 require (
-	github.com/garyburd/redigo v1.6.3
-	github.com/scottjbarr/yahoofinance v0.0.0-20170212024537-071cc11d69b5
+	github.com/gomodule/redigo v1.8.9
+	github.com/gorilla/mux v1.8.0
+	github.com/scottjbarr/config v0.0.4
+	github.com/scottjbarr/redis v0.0.4
 )
 
 require (
-	github.com/gomodule/redigo v1.8.9 // indirect
-	github.com/gorilla/mux v1.8.0 // indirect
-	github.com/scottjbarr/redis v0.0.4 // indirect
+	github.com/aws/aws-sdk-go v1.36.12 // indirect
+	github.com/google/uuid v1.1.2 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	github.com/kelseyhightower/envconfig v1.4.0 // indirect
+	github.com/scottjbarr/middleware v0.0.1 // indirect
+	github.com/scottjbarr/storage v0.0.1 // indirect
+	go.uber.org/atomic v1.6.0 // indirect
+	go.uber.org/multierr v1.5.0 // indirect
+	go.uber.org/zap v1.16.0 // indirect
 )
