+++ pkg/broadcaster/broadcaster.go
@@ -1,24 +1,24 @@
-package pricebroadcaster
+package broadcaster
 
 import (
 	"encoding/json"
 	"fmt"
 
-	redigo "github.com/gomodule/redigo/redis"
 	"github.com/scottjbarr/pricebroadcaster/pkg/cache"
 	"github.com/scottjbarr/pricebroadcaster/pkg/models"
+	"github.com/scottjbarr/pricebroadcaster/pkg/publisher"
 )
 
 // Broadcaster is responsible for receiving and broadcasting
 // data
 type Broadcaster struct {
 	Room      string
-	Publisher Publisher
+	Publisher publisher.Publisher
 	Cache     cache.Cache
 }
 
 // New returns a new Broadcaster with the given Config
-func New(room string, pub Publisher, cache cache.Cache) (*Broadcaster, error) {
+func New(room string, pub publisher.Publisher, cache cache.Cache) (*Broadcaster, error) {
 	return &Broadcaster{
 		Room:      room,
 		Publisher: pub,
@@ -51,33 +51,4 @@ func (b Broadcaster) publish(ohlc *models.OHLC) error {
 	}
 
 	return b.Cache.Set(ohlc.Symbol, data)
-}
-
-type Publisher interface {
-	Publish(string, []byte) error
-}
-
-type RedisPublisher struct {
-	Pool *redigo.Pool
-}
-
-func NewRedisPublisher(pool *redigo.Pool) *RedisPublisher {
-	return &RedisPublisher{
-		Pool: pool,
-	}
-}
-
-func (p *RedisPublisher) Publish(key string, b []byte) error {
-	conn := p.Pool.Get()
-	defer conn.Close()
-
-	if _, err := conn.Do("PUBLISH", key, b); err != nil {
-		return err
-	}
-
-	if err := conn.Flush(); err != nil {
-		return err
-	}
-
-	return nil
 }
